refactor selection handler
