+++ src/components/Cms/MdxEditor/helpers/lexical/select-end-of-div.ts
@@ -1,0 +1,17 @@
+export const selectEndOfDiv = (div: HTMLDivElement | null) => {
+    if (!div) {
+        return;
+    }
+    const range = document.createRange();
+    const selection = window.getSelection();
+
+    // Set range to end of div
+    range.selectNodeContents(div);
+    range.collapse(false); // false = collapse to end
+
+    // Apply the selection
+    selection?.removeAllRanges();
+    selection?.addRange(range);
+
+    div.focus();
+};
